+++ src/Exception/ConfigurationException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Paysera\Component\RestClientCommon\Exception;
+
+use Exception;
+
+class ConfigurationException extends Exception
+{
+
+}